accomplish; debugging

+++ fedml_api/data_preprocessing/spectrum/data_loader.py
@@ -75,7 +75,7 @@ def batch_data(data, batch_size):
         batched_x = data_x[i:i + batch_size]
         batched_y = data_y[i:i + batch_size]
         batched_x = torch.from_numpy(np.asarray(batched_x)).float()
-        batched_y = torch.from_numpy(np.asarray(batched_y)).long()
+        batched_y = torch.from_numpy(np.asarray(batched_y)).float()
         batch_data.append((batched_x, batched_y))
     return batch_data
